pyproject.toml
[tool.poetry]
name = "project-dz"
version = "0.1.0"
description = ""
authors = ["SvAS"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


